Removes unneeded packages and adds install support
I was a little overzealous in adding in all of the executables
from GAMP. We actually only needed gamp, vamp, and hgamp, so
since the 'fit' utility was giving me such a hard time to get it
compiled, and the fact it wasn't needed, I've removed it from the
package entirely along with logview.

I also added install support so that the user can use
'make install' to install the binaries into whatever prefix
they specified.