Fix name collision error
The particle class has a method in it that collided with the Decay's
constructor. I had to rename all the functions related to these
operations in order to break the name collision. There is still plenty
of work to be done though- The entire project needs to be updated to the
latest C++ standards, and namespaces need to be introduced to properly
prevent future namespace collisions.

Also, the Bison file that produced the KeyParser needs to be found
somewhere, it currently has a mountain of warnings that could become
errors in future versions of C++.