Remove time-bombs from particle
I removed a few time-bombs from the Decay library, and I also set
some variables to sane defaults that could potentially be used without
being initialized first.

Those time-bombs though: In a handful of locations the decay would wrap
it's functionality with `if (this)`. This is bad. This is really really
bad. At best, this is extra safety thrown in there but a naive developer,
at worst (and my fear) a solution to really bad design. Typically, `this`
should always be true, but if, somehow, the function is called when Decay
is uninitialized, `this` would be false.

This is bad design. If `this` is false in any situation, then it's surely
a bug that must be fixed.

I am hoping that this was just an extra bit of safety, becuase if it is
not, I just exposed a mountain of design bugs that **must** be fixed.