Remove all erroring code in PPGEN
Most of the issues in PPGEN was due to code that was called, but the
results were never used. As far as I'm aware, none of the code I
removed could have had implicit side effects, but there is still
a ton of questionable code in the repository.

I have two worries about the code in it's current state, potential
side effects from output changing, and potential side effects from
removing some calls to the random number generator.

I'm not sure if I can trust that this code base does not have some
weird assumptions about random numbers and their not really random
nature.